+++ CHANGELOG.md
@@ -30,6 +30,13 @@ and this project adheres to a modified version of [Semantic Versioning](./README
 - chore: Update NPM dependencies.
 - ci: Update workflow actions to latest versions.
 - ci: Fix XDebug install in Docker for PHP 7.x.
+- ci: Add `INCLUDE_EXTENSIONS` env variable for running 3rd-party plugin tests.
+- tests: Refactor helper methods.
+- tests: refactor functional `Cept` tests to `Cest` format.
+- tests: Add test for `NONE` provider enum.
+- tests: Add test for `Utils::is_current_user()` with empty user supplied.
+- tests: Backfill case for `TypeRegistry::get_registered_types()`.
+- tests: Add acceptance test for login with WC enabled.
 
 ## [0.0.3] - 2022-12-03